requirement.txt
absl-py==1.4.0
addict==2.4.0
boto3==1.26.123
botocore==1.29.123
brotlipy==0.7.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorlog==6.7.0
contextlib2==21.6.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
easydict==1.10
einops==0.6.1
filelock==3.9.0
fonttools==4.39.3
fsspec==2023.5.0
ftfy==6.1.1
gmpy2==2.1.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.8.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
mpmath==1.2.1
networkx==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
pip==23.0.1
platformdirs==3.10.0
prettytable==3.8.0
protobuf==4.24.1
psutil==2.1.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-transformers==1.2.0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.98
setproctitle==1.3.2
setuptools==66.0.0
six==1.16.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
timm==0.9.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.7
wheel==0.38.4
yapf==0.40.1
zhon==1.1.5
zipp==3.15.0
